style: Clean up
